+++ BALSAMIC/utils/cli.py
@@ -256,6 +256,8 @@ def get_snakefile(analysis_type, sequencing_type="targeted"):
     snakefile = Path(p, "workflows", "balsamic.smk")
     if analysis_type == "generate_ref":
         snakefile = Path(p, 'workflows', 'reference.smk')
+    if analysis_type == "pon":
+        snakefile = Path(p, 'workflows', 'PON.smk')
     elif analysis_type == "umi":
         snakefile = Path(p, "workflows", "UMIworkflow.smk")
 
@@ -609,3 +611,14 @@ def job_id_dump_to_yaml(job_id_dump: Path, job_id_yaml: Path, case_name: str):
                                                   "w") as jobid_out:
         jobid_list = jobid_in.read().splitlines()
         yaml.dump({case_name: jobid_list}, jobid_out)
+
+
+def create_pon_fastq_symlink(pon_fastqs, symlink_dir):
+    for fastq_name in os.listdir(pon_fastqs):
+        pon_fastq = Path(pon_fastqs, fastq_name).as_posix()
+        pon_sym_file = Path(symlink_dir, fastq_name).as_posix()
+        try:
+            LOG.info(f"Creating symlink {fastq_name} -> {pon_sym_file}")
+            os.symlink(pon_fastq, pon_sym_file)
+        except FileExistsError:
+            LOG.info(f"File {pon_sym_file} exists, skipping")